pyproject.toml
[tool.poetry]
name = "llm-realtor"
version = "0.1.0"
description = ""
authors = ["taejunsong <songtj10@snu.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
nltk = "^3.9.1"
transformers = "^4.49.0"
sentencepiece = "^0.2.0"
spacy = "^3.8.4"
numpy = "^2.2.3"
scikit-learn = "^1.6.1"
torch = "2.2.2"
torchtext = "0.17.2"
streamlit = "^1.43.2"
langchain-ollama = "^0.3.0"
langchain-community = "^0.3.20"
langchain-core = "^0.3.47"
langgraph = "^0.3.18"
langchain-text-splitters = "^0.3.7"
langsmith = "^0.3.18"
langchain = "^0.3.21"
python-decouple = "^3.8"
duckduckgo-search = "^2025.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
